rename styles to scss
