+++ web/components/Cards/CardEquipment.module.scss
