+++ web/components/Cards/CardNews.module.scss
