+++ web/components/Cards/CardNewsHorizontal.js
@@ -2,7 +2,7 @@ import { format, parseISO } from 'date-fns';
 import { es } from 'date-fns/locale';
 import Link from 'next/link';
 import CoverImage from '../coverImage';
-import styles from './CardNewsHorizontal.module.css';
+import styles from './CardNewsHorizontal.module.scss';
 
 const getDate = (date) => {
     const day = '0' + format(parseISO(date), 'd', { locale: es });
