+++ web/components/Cards/CardNewsHorizontal.module.scss
