--- web/components/Heros/HeroPage.module.css
@@ -1,3 +1,0 @@
-.hero {
-  @apply relative;
-}
