+++ web/components/Heros/HeroPage.module.scss
@@ -1,0 +1,3 @@
+.hero {
+    @apply relative;
+}
